backend/requirements.txt
django==3.2
gunicorn==20.1.0
django-heroku==0.3.1
djangorestframework==3.12.4
django-cors-headers==3.7.0
djoser==2.1.0
validate-email==1.3
psycopg2==2.8.6
whitenoise==5.2.0
dj-database-url==0.5.0